// src/redux/enums/ProductoEnum.ts
export enum ProductoEnum {
   POPULATE_PRODUCTO_START = 'POPULATE_PRODUCTO_START',
   POPULATE_PRODUCTO_SUCCESS = 'POPULATE_PRODUCTO_SUCCESS',
   POPULATE_PRODUCTO_ERROR = 'POPULATE_PRODUCTO_ERROR',

   ADD_PRODUCTO_START = 'ADD_PRODUCTO_START',
   ADD_PRODUCTO_SUCCESS = 'ADD_PRODUCTO_SUCCESS',
   ADD_PRODUCTO_ERROR = 'ADD_PRODUCTO_ERROR',

   DELETE_PRODUCTO_START = 'DELETE_PRODUCTO_START',
   DELETE_PRODUCTO_SUCCESS = 'DELETE_PRODUCTO_SUCCESS',
   DELETE_PRODUCTO_ERROR = 'DELETE_PRODUCTO_ERROR'
}